refactor: for node runtime handle DB migration at application startup (#174)
## Description

When running the container application in an orchestrated production
environment (Kubernetes) we should handle the migration at the startup
of the application, instead of having it as a separate scripts.

When bundling NextJS as standalone, we want to only include the minimal
dependencies in the final container image (see
https://github.com/better-chatbot/better-chatbot/pull/170) therefore
having everything inside the application startup makes sense.

## Related issues

Fixes https://github.com/better-chatbot/better-chatbot/issues/171

---------

Signed-off-by: axel7083 <42176370+axel7083@users.noreply.github.com>
Co-authored-by: choi sung keun <86150470+cgoinglove@users.noreply.github.com>