refactor(Dockerfile): follow nextjs recommendation (#170)
## Description

I was trying to deploy the
[better-chatbot](https://github.com/cgoinglove/better-chatbot/pkgs/container/better-chatbot)
container image to my Kubernetes cluster, and faced several issues.

This container image seems to be build and deploy by
[workflows/container.yaml](https://github.com/cgoinglove/better-chatbot/actions/workflows/container.yaml).


https://github.com/cgoinglove/better-chatbot/blob/b41cb200e4ba8b605f4a9bba13631513cfefdbcc/.github/workflows/container.yaml#L38-L46

--- 

The error I am getting is the following

```
 ERROR  EACCES: permission denied, mkdir '/.local/share/pnpm/.tools/pnpm/10.2.1_tmp_8'
For help, run: pnpm help run
```

And while trying to better understand this issue, I was reading the
[nextjs.org/docs/pages/getting-started/deploying#docker](https://nextjs.org/docs/pages/getting-started/deploying#docker)
and some element feel strange

1. Why using bun? I am not finding lot of info about nextJS and Bun as
officially supported, this may cause issues


https://github.com/cgoinglove/better-chatbot/blob/b41cb200e4ba8b605f4a9bba13631513cfefdbcc/docker/Dockerfile#L20-L22

2. Why `uv` is installed ?


https://github.com/cgoinglove/better-chatbot/blob/b41cb200e4ba8b605f4a9bba13631513cfefdbcc/docker/Dockerfile#L24-L25

From what I can read, the `https://github.com/astral-sh/uv` is
installed, but since the image is a node base, I am not sure this would
have the indented goal?

4. `CMD` seems odd ?
 

https://github.com/cgoinglove/better-chatbot/blob/b41cb200e4ba8b605f4a9bba13631513cfefdbcc/docker/Dockerfile#L42

In the nextJS documentation for deploying as a Docker image, they
recommend to configure the output as `standalone` and run `next build`.
They provide an example of a `Dockerfile`
(https://github.com/vercel/next.js/blob/canary/examples/with-docker/Dockerfile).

## Testing

Using podman

```
$: podman build -t localhost/better-chatbot:latest -f .\docker\Dockerfile .
$: podman run -it --env=POSTGRES_URL="<url>" localhost/better-chatbot:latest
```

Signed-off-by: axel7083 <42176370+axel7083@users.noreply.github.com>
Co-authored-by: choi sung keun <86150470+cgoinglove@users.noreply.github.com>